chore: 프로젝트 그룹 분리, APIManager.swift파일내 registProduct메서드 수정
