--- OpenMarket/OpenMarket/Model/MimeType.swift
@@ -1,12 +1,0 @@
-//
-//  MimeType.swift
-//  OpenMarket
-//
-//  Created by 이예원 on 2021/08/20.
-//
-
-import Foundation
-
-enum MimeType: String {
-    case png = "image/png"
-}
