+++ OpenMarket/OpenMarket/Resources/AppDelegate.swift
