+++ OpenMarket/OpenMarket/Views/Assets.xcassets/AccentColor.colorset/Contents.json
