+++ OpenMarket/OpenMarket/Views/Assets.xcassets/AppIcon.appiconset/Contents.json
